backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sixman"]

[tool.poetry.dependencies]
python = "^3.8.2"
Django = "3.2.6"
flake8 = "^4.0.1"
isort = "^5.10.1"
djangorestframework = "^3.13.1"
mysqlclient = "^2.1.0"
django-cors-headers = "^3.10.1"
drf-yasg = "^1.20.0"
djangorestframework-jwt = "^1.11.0"
gunicorn = "^20.1.0"
black = "^21.12b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
